+++ crates/stackable-operator/src/commons/authentication/ldap.rs
@@ -11,6 +11,7 @@ use crate::{
             tls::{TlsClientDetails, TlsClientDetailsError},
             SECRET_BASE_PATH,
         },
+        networking::Host,
         secret_class::{SecretClassVolume, SecretClassVolumeError},
     },
 };
@@ -36,8 +37,8 @@ pub enum Error {
 )]
 #[serde(rename_all = "camelCase")]
 pub struct AuthenticationProvider {
-    /// Hostname of the LDAP server, for example: `my.ldap.server`.
-    pub hostname: String,
+    /// Host of the LDAP server, for example: `my.ldap.server`.
+    pub hostname: Host,
 
     /// Port of the LDAP server. If TLS is used defaults to 636 otherwise to 389.
     port: Option<u16>,